+++ include/onews.h
@@ -132,7 +132,7 @@ public:
 	char  type;           // news type   // type may be > NEWS_TYPE_NUM, for indicating that the news has been displayed in the stock window, do display it on the newspaper again
 	char  news_type()     { return type%NEWS_TYPE_NUM; }
 
-	long  news_date;           	 // date of the news
+	int32_t  news_date;           	 // date of the news
 
 	char  nation_color1;     // nation color, can't use nation_recno directly, because it may bankrupt one day
 	char  nation_color2;
